V2 vertex parsing (#5)
* Changes to vertex import/export [WIP Export Breaks]

* Fixed exports

* Removed unnecessary function

* Fixed bug with importing multiple-mesh vertices

* Added backwards compat for old broken UV1+

* Fixed bug where a skinned mesh with invalid weights could be exported with no relevant bones, and then not be reimported

* Fixed a bug where the exporter would not export vertices with separate loop data (UVs, color etc.) separately

* Moved classes around to make them more consistent, cleaned up commented out code

* Moved mesh preparation for skinned and unskinned export to common function, implemented auto-triangulation

* Replaced nonzero padding TypeError with a print

* Fixed crash when model has dupe face, also reduced chances of dupe face to 0
